Move query to msg, fix #3049
