+++ src/client.js
@@ -57,7 +57,6 @@ const inputs = [
 	"mode",
 	"nick",
 	"notice",
-	"query",
 	"quit",
 	"raw",
 	"topic",
